requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
pywebpush==2.0.1
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3